MedFusionFinder/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
Faker==9.8.1
elasticsearch==7.13.4
pysftp==0.2.9
transformers==4.9.1
torch==1.9.0
pandas==1.1.5
requests==2.25.1